// lib/graphql/generated/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigDecimal: { input: any; output: any; }
  BigInt: { input: any; output: any; }
  Bytes: { input: any; output: any; }
  /**
   * 8 bytes signed integer
   *
   */
  Int8: { input: any; output: any; }
  /**
   * A string representation of microseconds UNIX timestamp (16 digits)
   *
   */
  Timestamp: { input: any; output: any; }
};
// ...
export type EternalFarming = {
  __typename?: 'EternalFarming';
  bonusReward: Scalars['BigInt']['output'];
  bonusRewardRate: Scalars['BigInt']['output'];
  bonusRewardToken: Scalars['Bytes']['output'];
  id: Scalars['ID']['output'];
  isDeactivated?: Maybe<Scalars['Boolean']['output']>;
  minRangeLength: Scalars['BigInt']['output'];
  nonce: Scalars['BigInt']['output'];
  pool: Scalars['Bytes']['output'];
  reward: Scalars['BigInt']['output'];
  rewardRate: Scalars['BigInt']['output'];
  rewardToken: Scalars['Bytes']['output'];
  virtualPool: Scalars['Bytes']['output'];
};